HSEARCH-3359 Move bridges used in a single mapper test to that test class
